% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{CollectionTrawl}
\alias{CollectionTrawl}
\title{Creates a list of trawl functions  given a type of trawl and a vector of
times as peak times. Has the option to get primitives. Note that only
exponential trawl are implemented so far.}
\usage{
CollectionTrawl(times, params, type, prim = F)
}
\arguments{
\item{times}{Vector of times to evaluate the trawl function (or primitive)
at.}

\item{params}{List of trawl parameters.}

\item{type}{Trawl type (so far, only "exp" is available).}

\item{prim}{Boolean to use primitive or not. Default is False (F).}
}
\description{
Creates a list of trawl functions  given a type of trawl and a vector of
times as peak times. Has the option to get primitives. Note that only
exponential trawl are implemented so far.
}
