% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{ComputeAExp}
\alias{ComputeAExp}
\title{Wrapper to compute total area under exponential trawl function.}
\usage{
ComputeAExp(rho)
}
\arguments{
\item{rho}{Exponential trawl parameter (should be non-negative).}
}
\value{
Total area under exponential trawl function (1/rho).
}
\description{
Wrapper to compute total area under exponential trawl function.
}
\examples{
ComputeAExp(1) # should be 1.0
ComputeAExp(0.2) # should be 5.0.
}
