% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{ComputeBInterExp}
\alias{ComputeBInterExp}
\title{Wrapper to compute intersection area between \code{t1} and \code{t2}
under exponential trawl function.}
\usage{
ComputeBInterExp(rho, t1, t2)
}
\arguments{
\item{rho}{Exponential trawl parameter (should be non-negative).}

\item{t1}{First timestamp.}

\item{t2}{Second timestamp.}
}
\value{
Intersection area between \code{t1} and \code{t2}
under exponential trawl function.
}
\description{
Wrapper to compute intersection area between \code{t1} and \code{t2}
under exponential trawl function.
}
\examples{
ComputeBInterExp(1, t1 = 3, t2 = 5)
ComputeBInterExp(0.2, t1 = 7, t2 = 3)
}
