% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{DiffVal}
\alias{DiffVal}
\title{Centered differentiation calculation.}
\usage{
DiffVal(value.p, value.m, epsilon)
}
\arguments{
\item{value.p}{function value at \code{x+epsilon}.}

\item{value.m}{function value at \code{x-epsilon}.}

\item{epsilon}{offset value for \code{x}.}
}
\value{
\code{(value.p - value.m) / (2*epsilon)}.
}
\description{
Centered differentiation calculation.
}
\examples{
DiffVal(4, 2, 1)
DiffVal(2.5, 2, 2)
}
