% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{dlgpd}
\alias{dlgpd}
\title{Computes the probability density function of Generalised Pareto Distribution at x,
with shape parameter alpha and scale parameter beta.}
\usage{
dlgpd(x, alpha, beta)
}
\arguments{
\item{x}{value at which the pdf is evaluated.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter, should be positive.}
}
\value{
GPD pdf function evaluated at x with shape and scale parameters respectively alpha and beta.
}
\description{
Computes the probability density function of Generalised Pareto Distribution at x,
with shape parameter alpha and scale parameter beta.
}
\examples{
dlgpd(2.34, alpha = 1, beta = 2)

}
