% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{GammaBox}
\alias{GammaBox}
\title{Sample Gamma with small shape parameter \code{alpha*dx*dy}.}
\usage{
GammaBox(alpha, beta, dx, dy, n)
}
\arguments{
\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}

\item{dx}{x-axis multiplier.}

\item{dy}{y-axis multiplier.}

\item{n}{Sample size.}
}
\value{
n Gamma(alpha*dx*dy, beta) samples.
}
\description{
Sample Gamma with small shape parameter \code{alpha*dx*dy}.
}
