% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{GammaSqBox}
\alias{GammaSqBox}
\title{Wrapper for GammaBox using a square to multiply the shape parameter.}
\usage{
GammaSqBox(alpha, beta, dt, n)
}
\arguments{
\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}

\item{dt}{Both x-axis and y-axis multiplier.}

\item{n}{Sample size.}
}
\value{
n Gamma(alpha*dt*dt, beta) samples.
}
\description{
Wrapper for GammaBox using a square to multiply the shape parameter.
}
