% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation_latent_trawl.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Computes initial guess for Univariate Latent-Trawl model.
Uses \cdoe{fExtreme} package method \code{gpdFit} to fit MLE.}
\usage{
generate_parameters(data, cluster.size)
}
\arguments{
\item{data}{Exceedane timeseries to be used.}

\item{cluster.size}{Lag at which ACF is negligeable.}
}
\description{
Computes initial guess for Univariate Latent-Trawl model.
Uses \cdoe{fExtreme} package method \code{gpdFit} to fit MLE.
}
