% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation_latent_trawl.R
\name{GenerateParameters}
\alias{GenerateParameters}
\title{Computes initial guess for Univariate Latent-Trawl model.
Uses \code{fExtreme} package method \code{gpdFit} to fit MLE.}
\usage{
GenerateParameters(data, cluster.size)
}
\arguments{
\item{data}{Exceedane timeseries to be used.}

\item{cluster.size}{Lag at which ACF is negligeable.}
}
\description{
Computes initial guess for Univariate Latent-Trawl model.
Uses \code{fExtreme} package method \code{gpdFit} to fit MLE.
}
\examples{
# TODO ADD data
data(hourlyhourly_bloomsbury_air_pollution_2000_2017)
GenerateParameters(data = hourly_bloomsbury_air_pollution_2000_2017$O3[1:1000], cluster.size = 8)

}
