% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation_latent_trawl.R
\name{GetEstimateRho}
\alias{GetEstimateRho}
\title{Computes first-order estimate for \code{rho}.}
\usage{
GetEstimateRho(alpha, beta, kappa, cluster.size, data)
}
\arguments{
\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter. Must be positive.}

\item{kappa}{Exceedance probability parameter. Must be positive.}

\item{cluster.size}{Lag at which ACF is negligeable i.e. cluster size.}

\item{data}{Exceedane timeseries to be used.}
}
\description{
Computes first-order estimate for \code{rho}.
}
\examples{
# TODO ADD data
GetEstimateRho(alpha = 5, beta = 2, kappa = 3, cluster.size = 8, data = NA)

}
