% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{GradF}
\alias{GradF}
\title{Computes gradident of a function given parameters and offset value epsilon.}
\usage{
GradF(f, params, epsilon = 1e-06)
}
\arguments{
\item{f}{R function taking only a vector of parameter as input.}

\item{params}{Vector of parameters.}

\item{epsilon}{Offset value for all components.}
}
\value{
Gradient of function \code{f} at parameters \code{params} with offset \code{epsilon}.
}
\description{
Computes gradident of a function given parameters and offset value epsilon.
}
