% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{HessianF}
\alias{HessianF}
\title{Computes Hessian matrix of a function given parameters and offset value epsilon.}
\usage{
HessianF(f, params, epsilon = 1e-06)
}
\arguments{
\item{f}{R function taking only a vector of parameter as input.}

\item{params}{Vector of parameters.}

\item{epsilon}{Offset value for all components.}
}
\description{
Computes Hessian matrix of a function given parameters and offset value epsilon.
}
\examples{
HessianF(function(params){prod(params^2)}, params = c(0.5, 7), epsilon = 1e-6)
HessianF(function(params){sum(params^3)}, params = c(2.5, 7), epsilon = 1e-6)
}
