% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{MarginalSimpleLik}
\alias{MarginalSimpleLik}
\title{Simplified marginal (GPD) log-likelihood function under independence in the
exponential trawl case.}
\usage{
MarginalSimpleLik(values, params)
}
\arguments{
\item{values}{Exceedance values.}

\item{params}{(at least) 2-d vector for GPD parameters}
}
\value{
Log-likelihood of GPD distribued variables (i.e. non-zero exceedances).
}
\description{
Simplified marginal (GPD) log-likelihood function under independence in the
exponential trawl case.
}
\examples{
MarginalSimpleLik(c(2.0, 0.3, 6.15, 0, 0.31), c(2.1, 1.17, 0.52, 4.17)) # for GPD(2.1, 1.17)

}
