% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{pairwise_11_2_1}
\alias{pairwise_11_2_1}
\title{See PairwiseOneOne2 for description.}
\usage{
pairwise_11_2_1(t1, x1, t2, x2, alpha, beta, kappa, rho, B1, B2)
}
\description{
See PairwiseOneOne2 for description.
}
