% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{pairwise_11_2_4}
\alias{pairwise_11_2_4}
\title{See PairwiseOneOne2 for description.}
\usage{
pairwise_11_2_4(t1, x1, t2, x2, alpha, beta, kappa, rho, B2, B3)
}
\description{
See PairwiseOneOne2 for description.
}
