% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{pairwise_11_exp}
\alias{pairwise_11_exp}
\title{Computes latent trawl pairwise likelihood with \code{(x,x)} where \code{x > 0} and \code{y > 0}
with exponential trawl function.}
\usage{
pairwise_11_exp(t1, x1, t2, x2, alpha, beta, kappa, rho, transformation = F,
  n_moments = 4)
}
\arguments{
\item{t1}{First timestamp.}

\item{x1}{Positive value corresponding to \code{t1}.}

\item{t2}{Second timestamp.}

\item{x2}{Positive value corresponding to \code{t2}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{rho}{Exponential trawl parameter. Should be positive.}

\item{transformation}{Boolean to use the Marginal Transform (MT) method.}

\item{n_moments}{Number of moments achieved by transformed GPD marginals, if used.}
}
\value{
Latent trawl pairwise likelihood with \code{(x,x)} where \code{x > 0} and \code{y > 0}.
}
\description{
Computes latent trawl pairwise likelihood with \code{(x,x)} where \code{x > 0} and \code{y > 0}
with exponential trawl function.
}
