% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneOne12}
\alias{PairwiseOneOne12}
\title{Computes second part of first term in latent trawl pairwise likelihood with \code{(x,x)} where \code{x > 0} and \code{y > 0}.}
\usage{
PairwiseOneOne12(t1, x1, t2, x2, alpha, beta, kappa, rho, B2)
}
\arguments{
\item{t1}{First timestamp.}

\item{x1}{Positive value corresponding to \code{t1}.}

\item{t2}{Second timestamp.}

\item{x2}{Positive value corresponding to \code{t2}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{rho}{Exponential trawl parameter. Should be positive.}

\item{B2}{Intersection area between \code{t1} and \code{t2} (in this order).}
}
\value{
Second part of first term in latent trawl pairwise likelihood with \code{(x,x)} where \code{x > 0} and \code{y > 0}.
}
\description{
Computes second part of first term in latent trawl pairwise likelihood with \code{(x,x)} where \code{x > 0} and \code{y > 0}.
}
