% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneOne23}
\alias{PairwiseOneOne23}
\title{Computes second term in latent trawl pairwise likelihood with \code{(x,x)}
where \code{x > 0} and \code{y > 0}.}
\usage{
PairwiseOneOne23(x1, x2, alpha, beta, kappa, B2, trawlA)
}
\arguments{
\item{x1}{Positive value corresponding to \code{t1}.}

\item{x2}{Positive value corresponding to \code{t2}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{B2}{Intersection area between \code{t1} and \code{t2} (in this order).}

\item{trawlA}{Total trawl set measure / area under trawl function.}
}
\value{
Second term in latent trawl pairwise likelihood with \code{(x,x)}
  where \code{x > 0} and \code{y > 0}.
}
\description{
Computes second term in latent trawl pairwise likelihood with \code{(x,x)}
where \code{x > 0} and \code{y > 0}.
}
