% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneZero1}
\alias{PairwiseOneZero1}
\title{Computes first part of latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.}
\usage{
PairwiseOneZero1(x1, alpha, beta, kappa, trawlA)
}
\arguments{
\item{x1}{Positive value corresponding to \code{t1}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{trawlA}{Total trawl set measure / area under trawl function.}
}
\value{
first part of latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.
}
\description{
Computes first part of latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.
}
