% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneZero21}
\alias{PairwiseOneZero21}
\title{Computes partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x} positive.}
\usage{
PairwiseOneZero21(x1, alpha, beta, kappa, B1, trawlA)
}
\arguments{
\item{x1}{Positive value corresponding to \code{t1}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{B1}{Difference area between \code{t2} and \code{t1} (in this order).}

\item{trawlA}{Total trawl set measure / area under trawl function.}
}
\value{
Partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x} positive.
}
\description{
Computes partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x} positive.
}
