% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneZero22}
\alias{PairwiseOneZero22}
\title{Computes partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.}
\usage{
PairwiseOneZero22(x1, alpha, beta, kappa, B2, trawlA)
}
\arguments{
\item{x1}{Positive value corresponding to \code{t1}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{B2}{Intersection area between \code{t1} and \code{t2}.}

\item{trawlA}{Total trawl set measure / area under trawl function.}
}
\value{
Partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.
}
\description{
Computes partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.
}
