% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneZeroExp}
\alias{PairwiseOneZeroExp}
\title{Computes term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}
with exponential trawl function.}
\usage{
PairwiseOneZeroExp(t1, x1, t2, alpha, beta, kappa, rho, transformation = F,
  n_moments = 4)
}
\arguments{
\item{t1}{First timestamp.}

\item{x1}{Positive value corresponding to \code{t1}.}

\item{t2}{Second timestamp.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{rho}{Exponential trawl parameter. Should be positive.}

\item{transformation}{Boolean to use the Marginal Transform (MT) method.}

\item{n_moments}{Number of moments achieved by transformed GPD marginals, if used.}
}
\value{
Second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}
with exponential trawl function.
}
\description{
Computes term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}
with exponential trawl function.
}
