% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseZeroZero1}
\alias{PairwiseZeroZero1}
\title{Computes partial part of latent trawl pairwise likelihood with \code{(x,y) = (0,0)}.}
\usage{
PairwiseZeroZero1(alpha, beta, kappa)
}
\arguments{
\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}
}
\value{
Partial part of latent trawl pairwise likelihood with \code{(x,y) = (0,0)}.
}
\description{
Computes partial part of latent trawl pairwise likelihood with \code{(x,y) = (0,0)}.
}
