% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseZeroZero2}
\alias{PairwiseZeroZero2}
\title{Computes second part of latent trawl pairwise likelihood with \code{(x,y) = (0,0)}.}
\usage{
PairwiseZeroZero2(alpha, beta, kappa, B1, B2, B3)
}
\arguments{
\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{B1}{Difference area between times \code{t1} and \code{t2}.}

\item{B2}{Intersection area between times \code{t1} and \code{t2}.}

\item{B3}{Difference area between times \code{t2} and \code{t1}.}
}
\value{
Second part of latent trawl pairwise likelihood with \code{(x,y) = (0,0)}.
}
\description{
Computes second part of latent trawl pairwise likelihood with \code{(x,y) = (0,0)}.
}
