% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseZeroZeroExp}
\alias{PairwiseZeroZeroExp}
\title{Computes latent trawl pairwise likelihood with \code{(x,y) = (0,0)} for Exponential
Trawl funciton.}
\usage{
PairwiseZeroZeroExp(t1, t2, alpha, beta, kappa, rho)
}
\arguments{
\item{t1}{First timestamp.}

\item{t2}{Second timestamp.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{rho}{Exponential trawl parameter. Should be non-negative.}
}
\value{
Second part of latent trawl pairwise likelihood with \code{(x,y) = (0,0)}.
}
\description{
Computes latent trawl pairwise likelihood with \code{(x,y) = (0,0)} for Exponential
Trawl funciton.
}
