% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{plgpd}
\alias{plgpd}
\title{Computes the cumulative distribution function of Generalised Pareto Distribution at x,
with shape parameter alpha and scale parameter beta.}
\usage{
plgpd(x, alpha, beta, lower.tail = F)
}
\arguments{
\item{x}{value at which the pdf is evaluated.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter, should be positive.}
}
\value{
GPD CDF function evaluated at x with shape and scale parameters respectively alpha and beta.
}
\description{
Computes the cumulative distribution function of Generalised Pareto Distribution at x,
with shape parameter alpha and scale parameter beta.
}
