% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{rlexceed}
\alias{rlexceed}
\title{Simulation of extreme value path using latent trawl process. Transformed
marginals have scale parameter \code{1+kappa}.}
\usage{
rlexceed(alpha, beta, kappa, rho = NA, times, marg.dist, n,
  transformation, trawl.function = NA, trawl_fs = NA,
  trawl_fs_prim = NA, n_moments = 4, deep_cols = 30)
}
\arguments{
\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}

\item{kappa}{Additive constant to scale parameter \code{beta}.}

\item{rho}{Trawl parameters. Must be positive if Exponential trawls are used.}

\item{times}{Vectors of discret times.}

\item{marg.dist}{Name of infinitely divisible distribution for latent trawls.
Currenlty implemented: gamma, gaussian, generalised hyperbolic, generalised
inverse gaussian.}

\item{n}{Number of simulations (so far, only \code{n=1} is implemented).}

\item{transformation}{Boolean to apply marginal transform method. Default is
False (F).}

\item{trawl_fs}{collection of trawl functions indexed on \code{times}.}

\item{trawl_fs_prim}{collection of trawl functions primitives indexed on
\code{times}.}

\item{n_moments}{Number of finite moments for transformed marginals.}

\item{deep_cols}{Depth of reconstruction (columns). Default is 30.}
}
\description{
Simulation of extreme value path using latent trawl process. Transformed
marginals have scale parameter \code{1+kappa}.
}
