% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{SliceArea}
\alias{SliceArea}
\title{Computes the area of a slice for the Slice Partition method of trawl
functions.}
\usage{
SliceArea(i, j, times, trawl_f_prim)
}
\arguments{
\item{i}{main index.}

\item{j}{secondary index.}

\item{times}{vector of discret times at which the trawl function is
partionned.}

\item{trawl_f_prim}{Trawl function primitive.}
}
\description{
Computes the area of a slice for the Slice Partition method of trawl
functions.
}
