% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{TrawlSliceReconstruct}
\alias{TrawlSliceReconstruct}
\title{Performs trawl slices reconstuction to get gamma samples using the
independent measure scaterring.}
\usage{
TrawlSliceReconstruct(alpha, beta, times, marg.dist, n, trawl.fs,
  trawl.fs.prim, deep_cols = 30, ghyp.object = NA)
}
\arguments{
\item{alpha}{(Gamma) Shape parameter.}

\item{beta}{(Gamma) Scale parameter.}

\item{times}{Vectors of discret times.}

\item{marg.dist}{Name of infinitely divisible distribution. Currenlty
implemented: gamma, gaussian}

\item{n}{Number of simulations (so far, only \code{n=1} is implemented).}

\item{trawl.fs}{collection of trawl functions indexed on \code{times}.}

\item{trawl.fs.prim}{collection of trawl functions primitives indexed on
\code{times}.}

\item{deep_cols}{Depth of reconstruction (columns). Default is 30 and must be
positive.}

\item{ghyp.object}{Object from \code{ghyp} package when using GH or GIG
distributions.}
}
\value{
Samples using trawl slice reconstruction disributed using
  \code{marginal}.
}
\description{
Performs trawl slices reconstuction to get gamma samples using the
independent measure scaterring.
}
