% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{TrfFindOffsetScale}
\alias{TrfFindOffsetScale}
\title{Computes predefined transformed scale parameter.}
\usage{
TrfFindOffsetScale(alpha, beta, kappa, offset_shape)
}
\arguments{
\item{alpha}{Shape parameter of output data.}

\item{beta}{Part of scale parameter of output data (beta + kappa).
Should be positive.}

\item{kappa}{Part of scale parameter of output data.
Should be positive.}

\item{offset_shape}{Shape parameter of input data.}
}
\value{
Transformed scale parameter given original data parameters.
}
\description{
Computes predefined transformed scale parameter.
}
