% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{TrfJacobian}
\alias{TrfJacobian}
\title{Computes determinant of Jacobian of Marginal Transform (MT) method.}
\usage{
TrfJacobian(z, alpha, beta, kappa, offset_scale, offset_shape)
}
\arguments{
\item{z}{GPD(offset_shape, offset_scale)-distributed data.}

\item{alpha}{Shape parameter of output data.}

\item{beta}{Part of scale parameter of output data (beta + kappa).
Should be positive.}

\item{kappa}{Part of scale parameter of output data.
Should be positive.}

\item{offset_scale}{Scale parameter of input data.
Should be positive.}

\item{offset_shape}{Shape parameter of input data.}
}
\value{
Transformed scale parameter given original data parameters.
}
\description{
Computes determinant of Jacobian of Marginal Transform (MT) method.
}
