% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation_latent_trawl.R
\name{Zeta}
\alias{Zeta}
\title{Zeta function from auto-correlation expansion for small lags.}
\usage{
Zeta(alpha, beta, kappa)
}
\arguments{
\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter. Must be positive.}

\item{kappa}{Exceedance probability parameter. Must be positive.}
}
\description{
Zeta function from auto-correlation expansion for small lags.
}
\examples{
Zeta(alpha = 5, beta = 4, kappa = 3)
}
