% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{EvaluateF}
\alias{EvaluateF}
\title{Evalutae function at given parameters.}
\usage{
EvaluateF(f, params, epsilon)
}
\arguments{
\item{f}{R function taking only a vector of parameters as input.}

\item{params}{Vector of parameters.}

\item{epsilon}{Offset value to add to all component of \code{params}.}
}
\description{
Evalutae function at given parameters.
}
\examples{
EvaluateF(function(params){prod(params^2)}, params = c(0.5, 7), epsilon = 0)
EvaluateF(function(params){sum(params)}, params = c(2.5, 7), epsilon = 1)
}
