% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{FullPL}
\alias{FullPL}
\title{Computes latent trawl FULL pairwise likelihood depending with exponential
trawl function.}
\usage{
FullPL(times, values, alpha, beta, kappa, rho, delta, logscale = T,
  transformation = F, trawl.function = "exp")
}
\arguments{
\item{times}{Vector of timestamps.}

\item{values}{Vector of target values.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{rho}{Trawl parameter(s). For \code{trawl.function="exp"}, it should be
positive.}

\item{delta}{Maximum depth of pairwise likelihood blocks. Should be positive
natural integer.}

\item{logscale}{Boolean to use logscale (log-likelihood). Default \code{T}.}

\item{transformation}{Boolean to use the Marginal Transform (MT) method.}
}
\value{
Full latent trawl pairwise likelihood.
}
\description{
Computes latent trawl FULL pairwise likelihood depending with exponential
trawl function.
}
