% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{GradF}
\alias{GradF}
\title{Computes gradident of a function given parameters and offset value epsilon.}
\usage{
GradF(f, params, epsilon = 1e-06)
}
\arguments{
\item{f}{R function taking only a vector of parameter as input.}

\item{params}{Vector of parameters.}

\item{epsilon}{Offset value for all components.}
}
\description{
Computes gradident of a function given parameters and offset value epsilon.
}
\examples{
GradF(function(params){prod(params^2)}, params = c(0.5, 7), epsilon = 1e-6)
GradF(function(params){sum(params)}, params = c(2.5, 7), epsilon = 1e-6)
}
