% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiate.R
\name{HessianF}
\alias{HessianF}
\title{Computes Hessian matrix of a $C^2$ function given parameters and offset value epsilon}
\usage{
HessianF(f, params, epsilon = 1e-06)
}
\arguments{
\item{f}{R function taking only a vector of parameter as input.}

\item{params}{Vector of parameters.}

\item{epsilon}{Offset value for all components.}
}
\value{
Hessian matrix of function \code{f} at parameters \code{params} with offset \code{epsilon}.
}
\description{
Computes Hessian matrix of a $C^2$ function given parameters and offset value epsilon
}
