% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneOne}
\alias{PairwiseOneOne}
\title{Computes latent trawl pairwise likelihood with \code{(x,x)} where \code{x >
0} and \code{y > 0} with exponential trawl function.}
\usage{
PairwiseOneOne(x1, x2, alpha, beta, kappa, B1, B2, B3,
  transformation = F, n.moments = 4)
}
\arguments{
\item{x1}{Positive value corresponding to \code{t1}.}

\item{x2}{Positive value corresponding to \code{t2}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{B1}{Difference area between \code{t1} and \code{t2} (in this order).}

\item{B2}{Intersection area between \code{t1} and \code{t2} (in this order).}

\item{B3}{Difference area between \code{t2} and \code{t1} (in this order).}

\item{transformation}{Boolean to use the Marginal Transform (MT) method.}

\item{n.moments}{Number of moments achieved by transformed GPD marginals, if
used.}
}
\value{
Latent trawl pairwise likelihood with \code{(x,x)} where \code{x > 0}
  and \code{y > 0}.
}
\description{
Computes latent trawl pairwise likelihood with \code{(x,x)} where \code{x >
0} and \code{y > 0} with exponential trawl function.
}
\examples{
PairwiseOneOne(x1=0.5, x2=0.3, alpha=2, beta=3, B1 = 0.3, B2 = 0.7,
                         B3 = 0.3, kappa=3.5)

}
