% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{PairwiseOneZero21}
\alias{PairwiseOneZero21}
\title{Computes partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.}
\usage{
PairwiseOneZero21(t1, x1, t2, alpha, beta, kappa, rho, B1)
}
\arguments{
\item{t1}{First timestamp.}

\item{x1}{Positive value corresponding to \code{t1}.}

\item{t2}{Second timestamp.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{rho}{Exponential trawl parameter. Should be positive.}

\item{B1}{Difference area between \code{t2} and \code{t1} (in this order).}
}
\value{
Partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.
}
\description{
Computes partial part of second term in latent trawl pairwise likelihood with \code{(x,0)} where \code{x > 0}.
}
