% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_latent_trawl.R
\name{SinglePairPL}
\alias{SinglePairPL}
\title{Computes correct latent trawl SINGLE pairwise likelihood depending on the values of \code{(x1,x2)}
with exponential trawl function.}
\usage{
SinglePairPL(t1, x1, t2, x2, alpha, beta, kappa, rho, transformation = F)
}
\arguments{
\item{t1}{First timestamp.}

\item{x1}{Positive value corresponding to \code{t1}.}

\item{t2}{Second timestamp.}

\item{x2}{Positive value corresponding to \code{t2}.}

\item{alpha}{Shape parameter. Should be positive.}

\item{beta}{Latent Gamma scale parameter. Should be positive.}

\item{kappa}{Exceedance probability parameter. Should be positive.}

\item{rho}{Exponential trawl parameter. Should be positive.}

\item{transformation}{Boolean to use the Marginal Transform (MT) method.}
}
\value{
SINGLE latent trawl pairwise likelihood depending on \code{(x1,x2)}.
}
\description{
Computes correct latent trawl SINGLE pairwise likelihood depending on the values of \code{(x1,x2)}
with exponential trawl function.
}
