% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{SliceArea}
\alias{SliceArea}
\title{Computes the area of a slice for the Slice Partition method of trawl
functions.}
\usage{
SliceArea(i, j, times, trawl.f.prim)
}
\arguments{
\item{i}{main index.}

\item{j}{secondary index.}

\item{times}{vector of discret times at which the trawl function is
partionned.}

\item{trawl.f.prim}{Trawl function primitive.}
}
\value{
Area of slice \code{S(i,j)} in Slice Partition method for trawl
  functions.
}
\description{
Computes the area of a slice for the Slice Partition method of trawl
functions.
}
\examples{
SliceArea(1, 2, times=c(0, 1, 2, 3), TrawlExpPrimitive(t=2,
  rho=0.2))
}
