% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_latent_trawl.R
\name{TrawlSliceReconstruct}
\alias{TrawlSliceReconstruct}
\title{Performs trawl slices reconstuction to get gamma samples using the
independent measure scaterring.}
\usage{
TrawlSliceReconstruct(alpha, beta, times, marg.dist, n, trawl.fs,
  trawl.fs.prim, deep_cols = 30, ghyp.object = NA)
}
\arguments{
\item{alpha}{(Gamma) Shape parameter.}

\item{beta}{(Gamma) Scale parameter.}

\item{times}{Vectors of discret times.}

\item{marg.dist}{Name of infinitely divisible distribution. Currenlty
implemented: gamma, gaussian, generalised hyperbolic (ghyp), generalised
inverse gaussian (gig).}

\item{n}{Number of simulations (so far, only \code{n=1} is implemented).}

\item{trawl.fs}{collection of trawl functions indexed on \code{times}.}

\item{trawl.fs.prim}{collection of trawl functions primitives indexed on
\code{times}.}

\item{deep_cols}{Depth of reconstruction (columns). Default is 30.}
}
\value{
Samples using trawl slice reconstruction disributed using
  \code{marginal}.
}
\description{
Performs trawl slices reconstuction to get gamma samples using the
independent measure scaterring.
}
\examples{
fcts <- lapply(c(1,2,3,4), function(t) TrawlExp(t, rho=0.2)) prims
  <- lapply(c(1,2,3,4), function(t) TrawlExpPrimitive(t, rho=0.2))
  TrawlSliceReconstruct(alpha=3, beta=2, times=c(0, 1, 2, 3), marg.dist =
  "gamma", n=1, fcts, prims)
  @importFrom ghyp ghyp
  @importFrom stats rgamma rnorm
}
