% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation_latent_trawl.R
\name{zeta}
\alias{zeta}
\title{Zeta function from auto-correlation expansion for small lags.}
\usage{
zeta(alpha, beta, kappa)
}
\description{
Zeta function from auto-correlation expansion for small lags.
}
